+++ src/konvigius/CHANGELOG.md
@@ -10,12 +10,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
  Planned improvements for next release:
 
-- Update and refine docstrings
+- Update and refine more docstrings
 - Additional examples in README
+- Possible bump to release 0.1.0
 
 ---
 
-## [0.l.0] - 2025-10-25
+## [0.0.4] - 2025-11-05
+
+- Fixed bug: auto inverted boolean field did not change value when source-field
+             was modified to a new value
+- Fixed bug: examples/demo-files -> show_vars() -> inspect_vars()
+- Fixed bug: Config.from_dict() raised error during validation in case of min/max
+             validation. The method starts now a transaction/commit to postpone
+             validations.
+- Added and updates docstrings
+- Added more pytests
+- 100% test coverage (219 passed)
+- Modified pyproject.toml to point correct location of this file
+
+---
+
+## [0.0.3] - 2025-10-25
 - Many internal improvements, tests, high coverage
 - API stable
 - Docs not yet updated